requirements.txt
# ADynamics Dependencies
# Conditional Flow Matching for AD MRI Disease Progression Modeling

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Medical imaging
monai>=1.2.0
nibabel>=5.0.0
SimpleITK>=2.2.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional: Logging
tensorboard>=2.14.0
wandb>=0.15.0  # Optional, for experiment tracking

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
